Integrated auto mapper in the edit game logic. Added logic to create game reviews(still in progress).  Added logic to get genre name by genre id. Added logic to get the genre id in the game view model. Change the layout and CSS of the save and details game page. Added reviews service. Added repositories for reviews, developers and publishers.

+++ GameDatabase/Models/EditGameModel.cs
@@ -14,7 +14,7 @@ namespace GameDatabase.Models
 
         public string Genre { get; set; }
 
-        public int GenreId { get; set; }
+        public decimal GenreId { get; set; }
 
         public IEnumerable<SelectListItem> Genres { get; set; }
